Fix mixed tests
